Listo, seguridad implementada

+++ api/inventarios/medicamentos/eliminar_medicamento.php
@@ -1,9 +1,19 @@
 <?php
+ob_start();
+
+header('Content-Type: application/json; charset=utf-8');
 require_once '../../../auth/roles.php';
-
 requireRoles(['admin_super']);
+include '../../../config/db.php';
 
-include "../../../config/db.php";
+$basura = ob_get_clean();
+if ($basura) {
+    echo json_encode(['ok' => false, 'msg' => 'Error interno del servidor.'], JSON_UNESCAPED_UNICODE);
+    exit;
+}
+
+$action = $_GET['action'] ?? '';
+
 $id = intval($_GET['id']);
 $mysqli->query("DELETE FROM medicamentos WHERE id_medicamento = $id");
 echo $mysqli->affected_rows > 0 ? "ok" : "error";
